Release version (#11)
* make dhall generate orders infinitely

* release version
