+++ src/components/types/waiter/waiter.go
@@ -1,7 +1,10 @@
 package waiter
 
+import "sync"
+
 type Waiter struct {
 	ID     int
 	Name   string
 	IsBusy bool
+	Mutex  *sync.Mutex
 }
