+++ src/controllers/controller.go
@@ -4,7 +4,6 @@ import (
 	"dininghall/src/components/constants"
 	"dininghall/src/components/types/order"
 	"dininghall/src/services"
-	coreService "dininghall/src/services"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -16,27 +15,19 @@ import (
 
 type Delivery order.Delivery
 
-/* func placeOrder(c *gin.Context) {
-	c.SetCookie("id", uuid.New().String(), 10, "http://localhost:4006/order", "http://localhost:4006/order", false, true)
-	c.Redirect(http.StatusFound, "http://localhost:4006/order")
-} */
-
-var wg sync.WaitGroup
-
+var waitGroup sync.WaitGroup
 var ratingPoints = 0
 var ordersCount = 0
 
 func distributeOrder(c *gin.Context) {
-	defer wg.Done()
 	c.JSON(200, "DHall: Delivery received, distributing...");
 
 	time.Sleep(constants.WaiterPickUpOrderTime * time.Second)
 	var delivery Delivery;
-	jsonDataRaw, err := ioutil.ReadAll(c.Request.Body)
-	if err != nil {}
-	e := json.Unmarshal(jsonDataRaw, &delivery)
-	if e != nil {}
+	jsonDataRaw, _ := ioutil.ReadAll(c.Request.Body)
+	_ = json.Unmarshal(jsonDataRaw, &delivery)
 	deliveryTime := time.Now().UnixMilli()
+
 	fmt.Printf("POST delivery %s received, distributing...\n", delivery.OrderID)
 	fmt.Printf("Delivery details:\n\t%v\n", delivery)
 	fmt.Printf("Time delivered:\n\t%v\n", time.Now().UnixMilli())
@@ -46,48 +37,21 @@ func distributeOrder(c *gin.Context) {
 	ratingPoints += currentOrderRating
 	ordersCount++
 
-	fmt.Printf("Rating: %d stars\n\n", currentOrderRating)
-	coreService.FinishOrder(delivery.WaiterID)
+	services.FinishOrder(delivery.WaiterID, delivery.TableID, &waitGroup)
 	println("DONE > ==========================================================")
-}
-
-/* func test(c *gin.Context) {
-	services.GenerateOrders(constants.GeneratedOrdersCount)
-} */
-
-func simulateOrdersConsecutively(c *gin.Context) {
-	// TODO: awaiting => waitGroup
-
-	for i := 0; i < constants.GeneratedOrdersCount; i++ {
-		wg.Add(1)
-		go services.GenerateOrder(i)
-	}
-
-	fmt.Println("Main: Waiting for workers to finish")
-	wg.Wait()
+	fmt.Printf("Rating: %d stars\n\n", currentOrderRating)
 	fmt.Printf("Rating points: %d\nOrdersCount: %d\n", ratingPoints, ordersCount)
 	fmt.Printf("Average rating for all orders: %g \n\n\n", float32(ratingPoints) / float32(ordersCount))
 }
 
-func getOrderList(c *gin.Context) {
-	id := c.Query("id")
-	items := c.Query("items")
-	priority := c.Query("priority")
-	maxWait := c.Query("maxWait")
+func simulateOrdersConsecutively(c *gin.Context) {
+	services.SimulateOrdersConsecutively(&waitGroup);
+	waitGroup.Wait()
 
-	c.JSON(200, gin.H{
-		"id":       id,
-		"items":    items,
-		"priority": priority,
-		"maxWait":  maxWait,
-	})
+	fmt.Println("!!!For some reason all go routines finished executing!!!")
 }
 
 func SetupController(router *gin.Engine) {
-  coreService.InitCoreService()
-
 	router.GET("/", simulateOrdersConsecutively)
-
-	// router.GET("/test", test)
 	router.POST("/distribution", distributeOrder)
 }
