+++ src/main.go
@@ -4,9 +4,12 @@ import (
 	"dininghall/src/configs"
 	"dininghall/src/controllers"
 	"dininghall/src/services"
+	"sync"
 
 	"github.com/gin-gonic/gin"
 )
+
+var waitGroup sync.WaitGroup
 
 func main() {
 	gin.ForceConsoleColor()
@@ -15,6 +18,12 @@ func main() {
 	configs.SetupENV()
 	services.InitCoreService()
 	controllers.SetupController(router)
+
+	// _, err := http.Post(os.Getenv("DHALL_URL")+"/", "application/json", nil)
+	// if err != nil {}
+
+	// services.SimulateOrdersConsecutively(&waitGroup)
+	// controllers.SimulateOrdersConsecutively(nil)
 
 	router.Run(":4005")
 }
